requirements.txt
wheel
pandas
pillow
seaborn
numpy
../pyutils

